*Never* jump on `pause`-events
Without this, attaching to a ruby/debug remote session causes the
debugger to pause at locations which were not intended by the user.
(e.g. rspec internals when debugging a spec, when attaching to a spec
which just started to run)

This is due to a bug (or weird intended behaviour) in ruby/debug, which
is fixed in https://github.com/ruby/debug/pull/640